fix: minor fix on unnecessary ambiguous event

+++ pkg/workflow/controllers/serial_node_reconciler.go
@@ -201,28 +201,30 @@ func (it *SerialNodeReconciler) syncChildNodes(ctx context.Context, node v1alpha
 					// delete that related nodes with best-effort pattern
 					nodesToDelete := finishedChildNodes[index:]
 
-					var nodesToCleanup []string
-					for _, item := range nodesToDelete {
-						nodesToCleanup = append(nodesToCleanup, item.Name)
-					}
-					it.eventRecorder.Event(&node, recorder.RerunBySpecChanged{CleanedChildrenNode: nodesToCleanup})
-
-					for _, refToDelete := range nodesToDelete {
-						nodeToDelete := v1alpha1.WorkflowNode{}
-						err := it.kubeClient.Get(ctx, types.NamespacedName{
-							Namespace: node.Namespace,
-							Name:      refToDelete.Name,
-						}, &nodeToDelete)
-						if client.IgnoreNotFound(err) != nil {
-							it.logger.Error(err, "failed to fetch outdated child node",
-								"node", fmt.Sprintf("%s/%s", node.Namespace, node.Name),
-								"child node", fmt.Sprintf("%s/%s", node.Namespace, nodeToDelete.Name))
+					if len(nodesToDelete)>0{
+						var nodesToCleanup []string
+						for _, item := range nodesToDelete {
+							nodesToCleanup = append(nodesToCleanup, item.Name)
 						}
-						err = it.kubeClient.Delete(ctx, &nodeToDelete)
-						if client.IgnoreNotFound(err) != nil {
-							it.logger.Error(err, "failed to fetch outdated child node",
-								"node", fmt.Sprintf("%s/%s", node.Namespace, node.Name),
-								"child node", fmt.Sprintf("%s/%s", node.Namespace, nodeToDelete.Name))
+						it.eventRecorder.Event(&node, recorder.RerunBySpecChanged{CleanedChildrenNode: nodesToCleanup})
+
+						for _, refToDelete := range nodesToDelete {
+							nodeToDelete := v1alpha1.WorkflowNode{}
+							err := it.kubeClient.Get(ctx, types.NamespacedName{
+								Namespace: node.Namespace,
+								Name:      refToDelete.Name,
+							}, &nodeToDelete)
+							if client.IgnoreNotFound(err) != nil {
+								it.logger.Error(err, "failed to fetch outdated child node",
+									"node", fmt.Sprintf("%s/%s", node.Namespace, node.Name),
+									"child node", fmt.Sprintf("%s/%s", node.Namespace, nodeToDelete.Name))
+							}
+							err = it.kubeClient.Delete(ctx, &nodeToDelete)
+							if client.IgnoreNotFound(err) != nil {
+								it.logger.Error(err, "failed to fetch outdated child node",
+									"node", fmt.Sprintf("%s/%s", node.Namespace, node.Name),
+									"child node", fmt.Sprintf("%s/%s", node.Namespace, nodeToDelete.Name))
+							}
 						}
 					}
 					break
